bit of clean up

+++ server/src/main/java/org/elasticsearch/script/field/vectors/ByteKnnDenseVector.java
@@ -66,7 +66,7 @@ public class ByteKnnDenseVector implements DenseVector {
         int i = 0;
         int j = docVector.offset;
         while (i < docVector.length) {
-            result += docVector.bytes[j++] * (int) queryVector[i++];
+            result += docVector.bytes[j++] * queryVector[i++];
         }
         return result;
     }
